Routes - Using validations in routes

+++ routes/students.routes.js
@@ -1,7 +1,7 @@
 const { Router } = require('express')
 const router = Router()
 const Students = require('../controllers/students.controller')
-const { validate } = require('../middleware/students.middleware')
+const { validate, validateId, validateIdDB } = require('../middleware/students.middleware')
 
 router.get('/', async (req, res) => {
     try {
@@ -36,6 +36,11 @@ router.put('/:id', async (req, res) => {
     try {
         let id = req.params
         let {body} = req
+        let validId = validateId(id.id)
+        if(!validId) return res.status(400).send({status: 400, msg: 'Invalid id'})
+        let data = await Students.getStudents()
+        let valid_IDDB = validateIdDB(id.id, data)
+        if(!valid_IDDB) return res.status(400).send({status:400, msg: 'User ID does not exist'})
         let valid = validate(body)
         if(!valid) return res.status(400).send({status: 400, msg: 'Invalid data (name o lastname)'})
         await Students.putStudents(id, body)
@@ -52,6 +57,11 @@ router.put('/:id', async (req, res) => {
 router.delete('/:id', async (req, res) => {
     try {
         let id = req.params
+        let validId = validateId(id.id)
+        if(!validId) return res.status(400).send({status: 400, msg: 'Invalid id'})
+        let data = await Students.getStudents()
+        let valid_IDDB = validateIdDB(id.id, data)
+        if(!valid_IDDB) return res.status(400).send({status:400, msg: 'User ID does not exist'})
         await Students.deleteStudents(id)
         return res.status(200).send({
             status: 200,
